tests: Triage new Go repository tests added/modified in Go 1.8.
There were 3 failing tests.

One was a modified old test that wasn't previously run due to a build
constraint (that has been removed in 1.8). It fails due to a compiler
bug, see #551 that tracks it.

Two were new tests that use unsupported packages/funcs (such as "unsafe",
runtime.FuncForPC, runtime.Callers) and failed as a result.